requirements.txt
numpy>=1.24.0
scikit-learn>=1.0.2
python-dotenv>=0.19.0
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0 